Separate utils class into 2
